+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/TargetedSequencingService.java
@@ -1,17 +1,9 @@
 package uk.ac.bbsrc.tgac.miso.core.service;
 
-import java.io.IOException;
-import java.util.Collection;
-import java.util.List;
-
 import uk.ac.bbsrc.tgac.miso.core.data.impl.TargetedSequencing;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginatedDataSource;
 
-public interface TargetedSequencingService extends PaginatedDataSource<TargetedSequencing> {
+public interface TargetedSequencingService extends DeleterService<TargetedSequencing>, ListService<TargetedSequencing>,
+    SaveService<TargetedSequencing>, PaginatedDataSource<TargetedSequencing> {
 
-  TargetedSequencing get(long targetedSequencingId) throws IOException;
-
-  Collection<TargetedSequencing> list() throws IOException;
-
-  List<TargetedSequencing> list(List<Long> targetedSequencingIds) throws IOException;
 }