+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/TargetedSequencingStore.java
@@ -1,13 +1,14 @@
 package uk.ac.bbsrc.tgac.miso.persistence;
 
 import java.io.IOException;
-import java.util.List;
 
 import uk.ac.bbsrc.tgac.miso.core.data.impl.TargetedSequencing;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginatedDataSource;
 
-public interface TargetedSequencingStore extends Store<TargetedSequencing>, PaginatedDataSource<TargetedSequencing> {
+public interface TargetedSequencingStore extends SaveDao<TargetedSequencing>, PaginatedDataSource<TargetedSequencing> {
 
-  List<TargetedSequencing> list(List<Long> targetedSequencingIds) throws IOException;
+  public TargetedSequencing getByAlias(String alias) throws IOException;
+
+  public long getUsage(TargetedSequencing targetedSequencing) throws IOException;
 
 }