+++ sqlstore/src/main/resources/db/migration/V8501__tarseq_description_optional.sql
@@ -1,0 +1,1 @@
+ALTER TABLE TargetedSequencing MODIFY COLUMN description varchar(255);